添加部分接口jwt鉴权

+++ server/api/category/delete.post.ts
@@ -4,6 +4,14 @@ const prisma = new PrismaClient();
 
 export default defineEventHandler(async (event) => {
   try {
+    // 从请求头中获取 Authorization
+    const authHeader = getHeader(event, "authorization");
+    if (!authHeader) {
+      return useErrorWrapper("", 401, false, "Authorization header missing");
+    }
+    const token = authHeader.split(" ")[1];
+    verifyToken(token);
+
     const { id } = await readBody(event);
     // 删除对应标签
     await prisma.category.delete({
